chore: Unify `ProductInfoWithTaxSupport` into `ProductInfo`
