+++ src/stories/organisms/product-info.stories.svelte
@@ -9,6 +9,9 @@
     subscriptionOptionWithTrial,
     nonSubscriptionOption,
     consumableProduct,
+    priceBreakdownTaxDisabled,
+    priceBreakdownTaxInclusive,
+    priceBreakdownTaxExclusive,
   } from "../fixtures";
 
   const { Story } = defineMeta({
@@ -30,6 +33,7 @@
     productDetails: product,
     purchaseOption: subscriptionOption,
     showProductDescription: true,
+    priceBreakdown: priceBreakdownTaxDisabled,
   }}
 />
 
@@ -39,6 +43,7 @@
     productDetails: product,
     purchaseOption: subscriptionOptionWithTrial,
     showProductDescription: true,
+    priceBreakdown: priceBreakdownTaxDisabled,
   }}
 />
 
@@ -48,5 +53,36 @@
     productDetails: consumableProduct,
     purchaseOption: nonSubscriptionOption,
     showProductDescription: true,
+    priceBreakdown: priceBreakdownTaxDisabled,
+  }}
+/>
+
+<Story
+  name="Tax Inclusive"
+  args={{
+    productDetails: product,
+    purchaseOption: subscriptionOption,
+    showProductDescription: true,
+    priceBreakdown: priceBreakdownTaxInclusive,
+  }}
+/>
+
+<Story
+  name="Tax Exclusive"
+  args={{
+    productDetails: product,
+    purchaseOption: subscriptionOption,
+    showProductDescription: true,
+    priceBreakdown: priceBreakdownTaxExclusive,
+  }}
+/>
+
+<Story
+  name="Tax Exclusive with trial"
+  args={{
+    productDetails: product,
+    purchaseOption: subscriptionOptionWithTrial,
+    showProductDescription: true,
+    priceBreakdown: priceBreakdownTaxExclusive,
   }}
 />
