+++ src/ui/organisms/product-info.svelte
@@ -3,22 +3,35 @@
     type Product,
     type PurchaseOption,
     ProductType,
+    type SubscriptionOption,
   } from "../../entities/offerings";
+  import PricingDropdown from "../molecules/pricing-dropdown.svelte";
+  import PricingTable from "../molecules/pricing-table.svelte";
   import ProductInfoHeader from "../molecules/product-header.svelte";
-  import ProductInfoPricing from "../molecules/product-pricing.svelte";
+  import PricingSummary from "../molecules/pricing-summary.svelte";
+  import { type PriceBreakdown } from "../ui-types";
 
   export let productDetails: Product;
   export let purchaseOption: PurchaseOption;
   export let showProductDescription: boolean;
+  export let priceBreakdown: PriceBreakdown;
 
   const isSubscription =
     productDetails.productType === ProductType.Subscription;
+
+  const subscriptionOption = purchaseOption as SubscriptionOption;
+
+  const basePhase = subscriptionOption?.base;
+  const trialPhase = subscriptionOption?.trial;
 </script>
 
 <section>
   <div class="rcb-pricing-info" class:has-expanded-details={isSubscription}>
     <ProductInfoHeader {productDetails} {showProductDescription} />
-    <ProductInfoPricing {productDetails} {purchaseOption} />
+    <PricingSummary {priceBreakdown} {basePhase} {trialPhase} />
+    <PricingDropdown>
+      <PricingTable {priceBreakdown} {trialPhase} />
+    </PricingDropdown>
   </div>
 </section>
